reorganized some classes and files
