+++ ZombustersWindows/GameObjects/Avatar.cs
@@ -1,89 +1,12 @@
-﻿#region File Description
-//-----------------------------------------------------------------------------
-// Game1.cs
-//
-// Microsoft XNA Community Game Platform
-// Copyright (C) Microsoft Corporation. All rights reserved.
-//-----------------------------------------------------------------------------
-#endregion
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Audio;
-using Microsoft.Xna.Framework.Content;
-
 using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
-#if !WINDOWS_PHONE
-//using Microsoft.Xna.Framework.Storage;
-#endif
-using GameStateManagement;
 
 namespace ZombustersWindows
 {
-
-    public struct AvatarState
-    {
-        public Vector2 position;
-        public List<Vector4> bullets;
-        public List<ShotgunShell> shotgunbullets;
-        public ObjectStatus status;
-        public int score;
-        public int lives;
-        public List<int> ammo;
-        public int lifecounter;
-        public float deathTimeTotalSeconds;
-        public Color color;
-        public int currentgun;
-        //public bool FiredGun;
-    }
-
-    public class ShotgunShell
-    {
-        public List<Vector3> Pellet; //perdigon
-        public Vector2 Position, Direction, Velocity;
-        public float Angle;
-
-        static readonly float pelletsSpreadRadians = MathHelper.ToRadians(2.5f);
-        const float initialSpeed = 640f;
-        protected float radius = 1f;
-
-        public ShotgunShell(Vector2 position, Vector2 direction, float angle, float totalgameseconds)
-        {
-            // initialize the graphics data
-            this.Position = position;
-            this.Angle = angle;
-            //this.Angle = (float)Math.Acos(Vector2.Dot(Vector2.UnitY, direction));
-            //if (direction.X > 0f)
-            //{
-            //    this.Angle *= -1f;
-            //}
-
-            this.Velocity =  initialSpeed * direction;
-
-            // calculate the direction vectors for the second and third projectiles
-            float rotation = (float)Math.Acos(Vector2.Dot(new Vector2(0f, -1f),
-                direction));
-            rotation *= (Vector2.Dot(new Vector2(0f, -1f),
-                new Vector2(direction.Y, -direction.X)) > 0f) ? 1f : -1f;
-            Vector2 Direction2 = new Vector2(
-                 (float)Math.Sin(rotation - pelletsSpreadRadians),
-                -(float)Math.Cos(rotation - pelletsSpreadRadians));
-            Vector2 Direction3 = new Vector2(
-                 (float)Math.Sin(rotation + pelletsSpreadRadians),
-                -(float)Math.Cos(rotation + pelletsSpreadRadians));
-
-            Pellet = new List<Vector3>();
-            Pellet.Add(new Vector3(this.Position, totalgameseconds));
-            Pellet.Add(new Vector3(this.Position, totalgameseconds));
-            Pellet.Add(new Vector3(this.Position, totalgameseconds));
-        }
-    }
-
     public class Avatar
     {
-        // Network/Draw state information
         public Vector2 position;
         public List<Vector4> bullets;
         public List<ShotgunShell> shotgunbullets;
@@ -148,7 +71,6 @@ namespace ZombustersWindows
                
         public void Initialize(Viewport view)
         {
-            //PRUEBASS
             this.entity = new SteeringEntity();
             this.entity.Width = 28;
             this.entity.Height = 50;
