+++ ZombustersWindows/GameObjects/Enemies/EnemyType.cs
@@ -1,0 +1,19 @@
+﻿namespace ZombustersWindows
+{
+    public enum EnemyType
+    {
+        Basic,
+        Looper,
+        Seeker,
+        Zombie,
+        Boomer,
+        Hunter,
+        Tank,
+        Smoker,
+        Witch,
+        Flamer,
+        Charger,
+        Spitter,
+        Jockey
+    }
+}
